facilities.c: properly restore session cloaks
0124 232639 -!- test (test) [ilbelkyr@gateway/test/x-wvgwoklegibimcqp] has joined #services

Set a project cloak:
0124 233021 -!- test [ilbelkyr@gateway/test/x-wvgwoklegibimcqp] has quit [Changing host]
0124 233021 -!- test (test) [ilbelkyr@project/example] has joined #services

Remove cloak:
0124 233032 -!- test [ilbelkyr@project/example] has quit [Changing host]
0124 233032 -!- test (test) [ilbelkyr@gateway/test/session] has joined #services
0124 233032 -!- test [ilbelkyr@gateway/test/session] has quit [Changing host]
0124 233032 -!- test (test) [ilbelkyr@gateway/test/x-wvgwoklegibimcqp] has joined #services

This is still not perfect but a lot better than the previous behaviour,
where the project cloak would be restored instead.

We store the computed facility cloak in per-user metadata; user metadata
is not actually written to disk, unlike myuser or mychan metadata. This
is necessary due to the "random" cloaking type.