Add IRCCloud cloaking type
This cloaking type is entirely specific to the IRCCloud service. It
operates in a fashion similar to to the ident cloaking type, however the
assigned gateway cloak will not include the 'u' or 's', only the
letters "id" and the subscriber ID itself.

Both uid1234 and sid1234 will thus result in a gateway/.../id1234 cloak.